Updated project structure
